refactor(design-system): hide Emotion behind the package's own Theme/API surface
Emotion becomes a fully internal implementation detail of
@soroush.tech/design-system instead of a peer dependency every consumer had to
install and partially expose.

- themes.ts now owns Theme, ThemeComponents, and every scale as native
  interfaces instead of augmenting @emotion/react's Theme. The augmentation
  surface consumers extend moves to
  `declare module '@soroush.tech/design-system/themes'`, verified to survive
  tsdown's chunked d.ts output the same way augmenting an external module used
  to.
- New src/engine.ts is the single internal seam wrapping Emotion's
  ThemeProvider/ThemeContext/useTheme/CacheProvider/Global, retyped against the
  owned Theme. styled.ts's styled.div/styled.span/... tag shorthands get their
  own retyped StyledTags for the same reason (Emotion's own were typed against
  its now-empty Theme).
- @emotion/react and @emotion/styled move from peerDependencies to
  dependencies — consumers no longer install Emotion themselves.
- markdown drops its own Emotion peer/dev deps entirely, retargeting its
  theme.mermaid / theme.components augmentation to the new module.
- apps/web had 18 files importing @emotion/react/@emotion/styled directly
  (pre-existing drift from the documented barrel-import convention); all
  switched to @soroush.tech/design-system, and its own @emotion/* runtime deps
  are removed (@emotion/server is kept — SSR-only, not Context-bound, so no
  singleton risk).
- Docs updated to match: design-system.md, theming.md, customization.md,
  markdown.md, both READMEs.

Verified with lint, typecheck, and 100% coverage (the exact CI commands, both
vitest projects combined) on design-system and markdown, plus lint, typecheck,
100% coverage, and a production build on apps/web.
