+++ apps/web/src/common/Flicker/Flicker.tsx
@@ -1,5 +1,4 @@
-import { keyframes } from '@emotion/react'
-import { styled } from '@soroush.tech/design-system'
+import { keyframes, styled } from '@soroush.tech/design-system'
 import { Flex, type FlexProps } from '@soroush.tech/design-system/Flex'
 
 export type FlickerProps = FlexProps
