+++ apps/web/src/common/Header/Header.tsx
@@ -1,8 +1,7 @@
 import { useState } from 'react'
-import { useTheme } from '@emotion/react'
 import SunIcon from 'src/assets/sun.svg?react'
 import MoonIcon from 'src/assets/moon.svg?react'
-import { styled } from '@soroush.tech/design-system'
+import { styled, useTheme } from '@soroush.tech/design-system'
 import { AppBar, type AppBarPosition } from '@soroush.tech/design-system/AppBar'
 import { View } from '@soroush.tech/design-system/View'
 import { Flex } from '@soroush.tech/design-system/Flex'
