+++ apps/web/src/test/utils/wrapper.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from '@testing-library/react'
 import { renderHook } from '@testing-library/react'
 import { useQueryClient } from '@tanstack/react-query'
-import { useTheme } from '@emotion/react'
+import { useTheme } from '@soroush.tech/design-system'
 import { describe, it, expect } from 'vitest'
 import {
   queryClient,
