+++ packages/design-system/package.json
@@ -546,11 +546,11 @@
   },
   "dependencies": {
     "@emotion/cache": "^11.14.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
     "@soroush.tech/styled-system": "workspace:^"
   },
   "peerDependencies": {
-    "@emotion/react": "^11.14.0",
-    "@emotion/styled": "^11.14.1",
     "@tanstack/react-form": "^1.33.2",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
@@ -562,8 +562,6 @@
   },
   "devDependencies": {
     "@emotion/css": "^11.13.5",
-    "@emotion/react": "^11.14.0",
-    "@emotion/styled": "^11.14.1",
     "@soroush.tech/eslint-config": "workspace:*",
     "@storybook/react-vite": "^10.5.0",
     "@tanstack/react-form": "^1.33.2",
