+++ packages/design-system/src/Button/README.md
@@ -23,7 +23,7 @@ Default: `"contained"`, overridable via `theme.defaults.buttonVariant` or `theme
 The value union is augmentable — register new values on the `ButtonVariants` interface and style them through `theme.components.Button.variants`:
 
 ```ts
-declare module '@emotion/react' {
+declare module '@soroush.tech/design-system/themes' {
   interface ButtonVariants {
     dashed: true
   }
