+++ packages/design-system/src/hooks/withTheme.test.tsx
@@ -1,9 +1,8 @@
 import { describe, it, expect } from 'vitest'
 import { render, screen } from '@testing-library/react'
-import type { CSSObject } from '@emotion/react'
 import { ThemeProvider } from '../ThemeProvider'
 import { baseTheme } from '../themes'
-import type { Theme } from '../themes'
+import type { CSSObject, Theme } from '../themes'
 import { StylesConsumer } from './StylesConsumer'
 import { withStyles } from './withStyles'
 import { withTheme } from './withTheme'
