+++ packages/markdown/src/themeComponents.spec.tsx
@@ -3,7 +3,7 @@ import { describe, it, expect, afterEach } from 'vitest'
 import { render, cleanup, fireEvent, screen } from '@testing-library/react'
 import { ThemeProvider } from '@soroush.tech/design-system/ThemeProvider'
 import { createTheme, baseTheme } from '@soroush.tech/design-system/themes'
-import type { ThemeComponents } from '@emotion/react'
+import type { ThemeComponents } from '@soroush.tech/design-system/themes'
 import { CodeBlock } from './CodeBlock'
 import { Control } from './Control'
 import { Editor } from './Editor'
