+++ packages/markdown/type-tests/augmentation.ts
@@ -1,11 +1,11 @@
 // Type-level test: the markdown package's `theme.components` slots must reach
 // consumers through the PUBLISHED d.ts. Runs via `pnpm test:types` (builds dist
 // first, then typechecks this isolated program). Guards that importing the
-// package barrel pulls in the `declare module '@emotion/react'` augmentation —
-// i.e. it survives tsdown's d.ts chunking — so `theme.components.MarkdownPreview`
-// and friends type-check for npm consumers.
+// package barrel pulls in the `declare module '@soroush.tech/design-system/themes'`
+// augmentation — i.e. it survives tsdown's d.ts chunking — so
+// `theme.components.MarkdownPreview` and friends type-check for npm consumers.
 import '@soroush.tech/markdown'
-import type { Theme, ThemeComponents } from '@emotion/react'
+import type { Theme, ThemeComponents } from '@soroush.tech/design-system/themes'
 
 // The three slots the package registers resolve to configurable component slots.
 export const editor: ThemeComponents['MarkdownEditor'] = {
